FLoC event logging condition update: log only if the floc has changed
Previously, we log the floc when it's computed and has a valid value.
Change it to be: log whenever the status changes, i.e. valid->invalid,
invalid->valid, or valid->different-valid.

Why:
1) This makes it more natural to understand.
2) This allows the server to know when the floc is disabled. This is not
necessary as of now, because currently the only updating trigger is
the scheduled update, so if the server sees no event after 24 hours
since the last event, the server could infer that the floc is disabled.
However, if we introduce new updating trigger like history-delete, then
the server would have insufficient knowledge. Thus, switch to the new
logging condition.


Bug: 1062736
Change-Id: I8796c2af89a4de5c5f8831e1c00693224927fcd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2357613
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799294}