Filter out overflow text in Select-to-Speak
When converting nodes into NodeGroup, we check each word one by one. If the word is out of the bound of its parent, we replace the
word with equal length space characters. The resulted text have the same length of the original text, and TTS will ignore overflow texts as they are empty now.

Bug: 775659
Change-Id: Iec3a296344b1c522ec255d56bfcbf51a68d50f5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2342127
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Lei Shi <leileilei@google.com>
Cr-Commit-Position: refs/heads/master@{#799307}