WebXR: handle DOM overlay from OOPIF fullscreen
If entering WebXR mode using DOM Overlay when we're already in
fullscreen mode, and if we have a remote ancestor frame that's
in another process, the ancestor frame didn't get set to transparent
since fullscreen_controller ignores changes to fullscreen options.

As a workaround, exit fullscreen mode in this case before proceeding
with the session request.

Bug: 1101193
Change-Id: I36b9918717abf6c5d462b7c78d545ae53fc5391e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2358493
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Commit-Queue: Klaus Weidner <klausw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799401}