Switch GPU pixel tests to full screenshots
Switches several GPU pixel tests on Fuchsia to use the new full
screenshot capabilities in Telemetry instead of the version that only
captures the visible portion of the web contents. This fixes the tests
on Fuchsia, where the device screen is significantly smaller than
previously tested devices. This path is not used everywhere because
capturing the screenshot using fromSuface: True is less likely to catch
GPU issues.

Bug: 1115680
Change-Id: I69423b4995978c1455978b7d9bdc6e80192b6ccd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2355133
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799433}