Force full damage on main back buffer for overlays.
This is Windows only.

DWM won't consider a swap chain as an overlay candidate
unless it's updating at high enough fps with full damages.

So we need to force full damage when the actual damage is
closer to full damage in order to promote the main back
buffer as an overlay.

This CL also added the bit to allow viz to make sure
swap chain's buffers always have up to date pixels
inside or outside damage rects. This enables us to
force full damage.

BUG=1117185
TEST=manual
R=sunnyps@chromium.org,magchen@chromium.org

Change-Id: Idff7a90952c810a07992df0ace040be92b63e4d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2359730
Commit-Queue: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Jonathan Backer <backer@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799445}