feed v2: Add dcheck to ensure feedv1 hosts aren't created
See bug. I was unable to reproduce the error, it's likley due to
some experiment. I suspect it may be related to instant start,
as this line would result in creating the wrong feed if the
coordinator is made too soon:
https://source.chromium.org/chromium/chromium/src/+/master:chrome/android/feed/core/java/src/org/chromium/chrome/browser/feed/FeedSurfaceCoordinator.java;l=223

For now, just DCHECK if this happens.

Bug: 1117680
Change-Id: Iecc2690698b34a04151b62b684a6b8b955055fe4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2363809
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Commit-Queue: Dan H <harringtond@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799448}