[Android] Create a feedback module.
Currently HelpAndFeedback is not modularized, so a modularized fragment
is not able to use it due to circular dependency.

In this CL, create a feedback module and introduce two interfaces as a
workaround:
* HelpAndFeedbackLauncher, which is used to launch a help and feedback
page in an activity. HelpAndFeedback implements this interface.
* FragmentHelpAndFeedbackLauncher, which is used to inject
HelpAndFeedbackLauncher to a fragment. Any fragment that needs to
launch a help page implements this interface.

The safe_browsing module will depend on this new module to launch a
help page.

Bug: 1115758,1117343
Change-Id: I3d9d76f0ba87db26ae5bbe853f7b6cc38eb9353a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2359749
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Natalie Chouinard <chouinard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799475}