Android: Capture more tombstones in instrumentation tests
When chrome_public_smoke_test has a crash, our test runner does not
notice and so does not collect tombstones.

This changes the "are there tombstones" logic to look for a tombstone
message in logcat rather than looking to see if any tests crashed.

Bug: 1117540
Change-Id: Ic0594f0d8dcb2cb8c7280ade9d8d5133c890614f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2363075
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799490}