[IntentHandling] Launch share intent with specific activity name.
Currently, the specific activity name is not passed from ARC. This
causes an issue where if we try to share a file to an app that has
multiple activities that can handle the file, it will trigger the
disambiguity dialog. This CL pass the activity name with the activity
label from ARC to ensure launching the correct activity.

BUG=1092784
Test: Manually test with share jpeg file to facebook app.

Change-Id: I4290ddce72c80548d23171f020284c39c87ce688
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2360094
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Melissa Zhang <melzhang@chromium.org>
Reviewed-by: David Jacobo <djacobo@chromium.org>
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799523}