[Refactor] Add interface for LTHI -> Input
This CL adds an interface through which LayerTreeHostImpl and the
"graphics" side of the compositor can talk to the input handler. For
now, LTHI keeps the concrete object and still owns it but eventually
this will be moved out and LTHI will only have a reference to the
interface.

A followup will add an interface for the other direction so that the
input system doesn't have to have a concrete reference to LTHI.

Bug: 915926
Change-Id: Ieb4c0642e6674b6f09e8c7d8b6ff366cc8179bce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2359190
Reviewed-by: Xida Chen <xidachen@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799593}