[omnibox] Fix buried Pedals by moving from duplicates to surface match
This CL resolves the occasional case of disappearing Pedals. If a pedal
is attached to a match that then becomes a duplicate, the remaining
surface match will not have a pedal because only one is allowed in the
whole set. In such cases, moving the pedal from the duplicate to the
main surface match will preserve the pedal for presentation.

Bug: 1118126
Change-Id: Ie25d6634f4face69e102521510d983ea5a723b5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2363872
Auto-Submit: Orin Jaworski <orinj@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799671}