Enable CorbAllowlistAlsoAppliesToOorCors via fieldtrial...config.json.
The CL tweaks 1 remaining test that relied on the legacy/deprecated
behavior that allowed CORS-exempt cross-origin requests from content
scripts.  The CL changes the test request to same-origin.  This change
should be okay - the cross-origin aspect of the request was not
essential to the test, which mostly wanted to verify that the XHR
results in the "blinkRequestResource XMLHttpRequest"
|domExpectedActivity|.

Bug: 920638
Change-Id: I9d9f765b91c7524586ac1b3280ba063274dba0e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2359392
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799677}