Respect emulation scale for window segments
We should not be using the original root window segments when none
are passed to the emulator, since these are based off of the DIPs the
widget is occupying, which doesn't take into account the emulator
scale (zoom drop down menu).

Instead, just emulate a single segment that is the same as the
emulated widget size, which takes the scale into account when emulated
width/height are not specified.

Bug: 1119306

Change-Id: Ic6ab80517989309f697036da873af3448981ad16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2350763
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#799682}