[CORS-RFC1918] Wire content setting to feature.
This change introduces PrivateNetworkRequestPolicy, an enum describing
how the network service should treat private network requests. A value
for this enum is communicated by the browser to the network service
through ClientSecurityState. This replaces the direct feature check by
URLLoader, making it slightly easier to test URLLoader::OnConnected().

The feature check is moved one layer up, to content/. NavigationRequest
is now in charge of setting the PrivateNetworkRequestPolicy according
to the state of the feature flag and the stated intent of the content
embedder, as communicated through ContentBrowserClient.

By default, ContentBrowserClient allows no exceptions to the feature.
ChromeContentBrowserClient uses content settings to manage exceptions.

Bug: chromium:986744
Change-Id: Ic053b643426e70a4e88260e25709b95a3b843810
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339972
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Titouan Rigoudy <titouan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799711}