Access Context Auditing: Add StoragePartition deletion observer
To support the observation of deletion events which flow through the
StoragePartition exposed interface, the ability to register deletion
observers is introduced.

This functionality is then used by the Access Context Audit service to
remove records for origin storages which are deleted in this manner.

Bug: 1101675
Change-Id: I4123f42ba6607503a479f9b75b03fdfeaba023a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2320569
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/master@{#799713}