WebLayer contacts picker: Split Chrome logic out of PickerAdapter
PickerAdapter is split into a base PickerAdapter and a
ChromePickerAdapter. Inheritance is used instead of a delegate object
because ChromePickerAdapter overrides some Adapter methods.

The code that uses the signin component is left in ChromePickerAdapter,
even though it's componentized, as WebLayer doesn't currently use this
component. This code may also be componentized later, but the plan is
for WebLayer to initially skip the special handling for the owner
contact.

After this change, all of the contacts_picker package save
ChromePickerAdapter can be moved to //components.

Bug: 1016938
Change-Id: Iba554c2064e1e302a05455b8875b409a14c3b199
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2324762
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799722}