[Android][WebSignIn] Adjust UI for sign-in general error screen
This CL adjusts the UI for the sign-in general error screen in the web
sign-in flow and add strings needed for the screen. The order of the
view elements in account_picker_bottom_sheet_view.xml are rearranged to
make sure that the height of sign-in general error screen is the same as
the sign-in in progress screen.

Screenshot: https://crbug.com/1114589#c5
Bug: 1114589
Change-Id: I281c59eb7a64f7121cf1a8dcd989a32c37f6a1a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2362588
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Tanmoy Mollik <triploblastic@chromium.org>
Commit-Queue: Alice Wang <aliceywang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799765}