Show lock URL and the origin-keyed bit in chrome://process-internals
After https://crrev.com/c/2304537, the site URL will no longer contain
the process lock URL in the hash, so the lock URL won't be visible on
chrome://process-internals.  This CL adds the plumbing to show the
lock URL independently from site URL, but only if the two URLs are
different.  This way, we'll be able to view the lock URL
in cases where it matters (e.g., for hosted apps), and it won't
pollute the UI in the common case where it matches the site URL.

For completeness, this CL also plumbs the is-origin-keyed bit, so that
all current parts of SiteInfo (foundation of a future
SecurityPrincipal abstraction) are shown.  The origin-keyed bit is
currently only shown when opt-in origin isolation is in effect.

Bug: 1105994, 1067389
Change-Id: Ife20bcefc57aecee96e24a436ae89902e0837d20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2364816
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799787}