[Extensions] Fix calls to parent SetUp functions in test classes.
A number of SetUp overrides were calling the parent override
before doing the setup for that class. This doesn't work because
any code following the call to the parent override is executed
after the current test fixture executes.

The code in WebrtcLoggingPrivateApiTest to set up the extension_
data member was also moved to SetUpOnMainThread. Building the
extension in SetUp crashes because the global ExtensionsClient
instance is not yet created.

Bug: 1117287
Change-Id: Ifd82c782b7f012cd0225766030fac2574ef68937
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2361004
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799817}