Add CHECK to help determine source of crash
Add a CHECK before we break out of the loop in ComputeIndices.  This
has never happened in my testing, so this is just a guess.  If we do
get crashes, perhaps we can get better information about what's
happening.

This is important because if we do exit the loop early, we're not
telling ComputeOutput about it and the read indices may not be set up
correctly for that.  Also, update frames_processed to tell ComputeOutput
how many values were computed so ComputeOutput doesn't compute more
outputs than needed.

I think letting this bake for a week or so will be enough to gather
enough information if this is the case.

Bug: 1116104
Change-Id: Id0ab1d3ca801196fc49b878349f469cbd4f6e11a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2358815
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799827}