chromeos: Support PIN auth for in-session auth dialog
If the submitted password looks like PIN, try authenticate the user
with PIN. On failure, submit the password as normal password. This
fallback to password auth is because the user may type the PIN into
the password textfield (though admittedly this is a less valid use
case). This behavior is consistent with ScreenLocker.

Bug: b:156258540
Change-Id: I23bf097a35c6458e8f341ae105679d6d42a83c59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2364756
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799846}