[Passwords] Affiliation Service implementation.
This change covers the implementation of AffiliationService, the AffiliationFetcherDelegate interface implementation and unit tests.
It prefetches data with a help of AffiliationFetcher and provides an API to get change password urls for requested origin.
This change includes only preparing data for prefetch without processing the response.

Bug: 1108279
Change-Id: I3f47a3b1f7c0b99726b4107d484387a77b1345b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332604
Commit-Queue: Nela Kaczmarek <nelakaczmarek@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799852}