a11y: Activate magnifier while drag a window across displays.
The root window and root_location of the event are still based on the
original display while dragging a window across displays. This makes
the magnifier will not be activated in the display that the window has
been dragged to until the drag be released.

This cl updates the root window and root location during drag to make
sure the magnifier can be activated at corresponding display
immediately while drag a window across displays.


Bug: 1082392
Change-Id: Ie3feb7679c29e1e9c0b4f8811ae98bd0f170ade0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354997
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Min Chen <minch@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799883}