Add support to enable granular permissions in GetAuthToken
For unbundled consent to be available for extensions to use, Chrome
needs to be able to explicitly request for it in an API call body.
Additionally, extension developers should be able to toggle the use of
unbundled consent during the transition into using unbundled consent.

This CL adds both requirements. The changes include adding an optional
parameter 'enableGranularPermissions' in GetAuthToken for developers to
toggle unbundled consent and adjusting Chrome's API calls accordingly
depending on the value of this optional parameter.

Bug: 1100535
Change-Id: I5051ebe2b304a0a93f4e98aea447c0668e5e0a3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2350544
Commit-Queue: William Lin <williamlin@google.com>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799918}