Files app: Fix FileManagerJsTest to work without DOMv0 and HTML Imports
Change the generated HTML to include the HTML import polyfil and to not
include the scripts from <cr_elements> because they'll be included via
HTML imports by the elements that use them.

Change the test harness to detect if HTML Import polyfil is available
and wait for imports to finish before starting the test.

Bug: 1111393
Change-Id: I2a8d55bee39f8386e75183bcbb79fbbf15204d03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2362167
Commit-Queue: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799934}