Use correct UKM source ID in WebRequestProxyingURLLoaderFactory
Currently we get a UKM source ID from the frame in
WebRequestAPI::MaybeProxyURLLoaderFactory but it's too early - the frame
has not been fully committed yet, and the UKM source ID for the previous
URL is returned.

Get the UKM source ID in
WebRequestProxyingURLLoaderFactory::CreateLoaderAndStart, to work around
the problem.

Bug: 1119320
Change-Id: I1fdd8d301aad916b11eef50e521d7f42087ed41c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2362071
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Commit-Queue: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799948}