Skip overlay if there are more than one videos in a frame.
For Windows 10 only.
We only promote one YUV overlay even if there are more than one YUV
quads in a frame. We might not save power if the rest of YUV quads are
drawn in the Renderer by 3D hardware.
The previous overlay strategy that promotes the quad with the biggest
pixel size is removed.

Bug: 1117240
Change-Id: Id729e5d8af842682528c2dcb9d0e70e4f49fea66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2362084
Commit-Queue: Maggie Chen <magchen@chromium.org>
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799958}