Add screenshots to the holding space model
Introduces HoldingSpaceItem::Type enumeration with types that will be
supported in the holding space UI - pinned files, downloads and
screenshots. Also adds data that will needed to represent the items in
the UI (in addition to the existing text item representation)
*   the item's absolute file path
*   the file system URL of the file backing the item
*   the image representation of the item

Changes how the item ID is calculated - instead of a randomly generated
GUID, the ID is now deffered from the item type and file system URL.

Changes HoldingSpacePresenter to track items per type (the class is
still only used by a test, but should be eventually be hooked up with
the holding space tray bubble code).

Updates screenshot grabber to add the screenshot to the holding space
after creating the screenshot file and loading the screenshot bitmap.

BUG=1112014

Change-Id: Ibc13dfb7bc59fbe5079bd4bafeb354f438dc7f86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2359230
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: David Black <dmblack@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799964}